Moved Aliases utility from core.util to util
