+++ holoviews/core/util.py
@@ -347,27 +347,6 @@ def capitalize_unicode_name(s):
     return s[:index] + tail
 
 
-class Aliases(object):
-    """
-    Helper class useful for defining a set of alias tuples on a single object.
-
-    For instance, when defining a group or label with an alias, instead
-    of setting tuples in the constructor, you could use
-    ``aliases.water`` if you first define:
-
-    >>> aliases = Aliases(water='H_2O', glucose='C_6H_{12}O_6')
-    >>> aliases.water
-    ('water', 'H_2O')
-
-    This may be used to conveniently define aliases for groups, labels
-    or dimension names.
-    """
-    def __init__(self, **kwargs):
-        for k,v in kwargs.items():
-            setattr(self, k, (k,v))
-
-
-
 class sanitize_identifier_fn(param.ParameterizedFunction):
     """
     Sanitizes group/label values for use in AttrTree attribute
